moved class to .h
